backend/requirements.txt
asgiref==3.8.1
asttokens==2.4.1
attrs==24.2.0
backcall==0.2.0
beautifulsoup4==4.12.3
bleach==6.1.0
certifi==2024.8.30
charset-normalizer==3.4.0
colorama==0.4.6
decorator==5.1.1
defusedxml==0.7.1
dj-database-url==2.2.0
Django==5.1.2
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
docopt==0.6.2
executing==2.1.0
fastjsonschema==2.20.0
idna==3.10
ipython==8.12.3
jedi==0.19.1
Jinja2==3.1.4
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
jupyter_client==8.6.3
jupyter_core==5.7.2
jupyterlab_pygments==0.3.0
MarkupSafe==3.0.2
matplotlib-inline==0.1.7
mistune==3.0.2
nbclient==0.10.0
nbconvert==7.16.4
nbformat==5.10.4
packaging==24.1
pandocfilters==1.5.1
parso==0.8.4
pickleshare==0.7.5
pillow==11.0.0
pipreqs==0.5.0
platformdirs==4.3.6
prompt_toolkit==3.0.48
psycopg2==2.9.10
pure_eval==0.2.3
Pygments==2.18.0
PyJWT==2.9.0
python-dateutil==2.9.0.post0
pywin32==308
pyzmq==26.2.0
referencing==0.35.1
requests==2.32.3
rest-framework-simplejwt==0.0.2
rpds-py==0.20.0
six==1.16.0
soupsieve==2.6
sqlparse==0.5.1
stack-data==0.6.3
tinycss2==1.3.0
tornado==6.4.1
traitlets==5.14.3
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.2.3
wcwidth==0.2.13
webencodings==0.5.1
yarg==0.1.9
